# Swagger/oas3.yaml
openapi: 3.0.0
info:
    title: Treport
    version: v.0
paths:  
...
    /notifications:
        get:
            description: Get a list of notifications. (/notification?user={user})
            parameters:
                -   in: query
                    name: user
                    description: Filter the notification by a given user. (Query value must be of the type [ "/users/:id" ])
                    schema:
                        type: integer
            responses:
                "200":
                    description: Notification list sent. (Attribute users is of the type [ "/users/:id" ], attribute report is of the type "/reports/:id")
                    content:
                        application/json:
                            schema:
                                type: array
                                $ref: "#/components/schemas/Notification_GET"
                "400":
                    description: User id's format is incorrect.
                "404": 
                    description: Notification not found.
        post:
            description: Create new notification.
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            $ref: "#/components/schemas/Notification_POST"
 
            responses:
                "201":
                    description: New notification added
                    headers:
                        "Location":
                            description: Link to the newly created notification
                            schema:
                                type: string
                "400":
                    description: May be an error in the report quering or in the user quering
                "404": 
                    description: No user has been found or no report has been found
    /notifications/{id}:
        get:
            description: Get a notification with a given id.
            responses:
                "200":
                    description: Notification sent correctly. (Attribute users is of the type [ "/users/:id" ], attribute report is of the type "/reports/:id")
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Notification_GET"
                "400":
                    description: Error in notification quering.
                "404":
                    description: Notification not found.
        
        delete:
            description: Delete a notification with a given id.
            responses:
                "204":
                    description: Notification deleted
                "400":
                    description: ID not accepted
                "404":
                    description: Notification not found
components:
    schemas:
...
        Notification_GET:
            type: object
            properties:
                id:
                    type: integer
                title:
                    type: string
                content:
                    type: string
                report:
                    type: string
                users:
                    type: array
                    items: 
                        type: string
        Notification_POST:
            type: object
            properties:
                title:
                    type: string
                content:
                    type: string
                report:
                    type: string
                users:
                    type: array
                    items: 
                        type: string
...